fix portal type title when deviates from FTI id
The portal_type metadata in the catalog contains a string of the ID of
the type's FTI.
The ID may be different from the title and the title is usually
translated.

For properly translating the type, we now use the FTI title with and the
FTI i18n_domain with fallbacks for when the FTI is not found or has no
i18n_domain.