Fix statusmap view when being called on an inactive content.
If the context is inactive (expired or in the future) and the user does not have the permissions to access future or expired content, the context was not included in the statusmap which was leading to a confusing error message in combination with certain publisher constraints.

This is solved by adding the context to the list of objects if it is not there yet.